RA|added link to stock register
