+++ opensrp-path/src/main/assets/vaccine_type.json
@@ -1,0 +1,44 @@
+[
+  {
+    "doses": "20",
+    "name": "BCG",
+    "openmrs_parent_entity_id": "",
+    "openmrs_date_concept_id": "",
+    "openmrs_dose_concept_id": ""
+  },
+  {
+    "doses": "25",
+    "name": "OPV",
+    "openmrs_parent_entity_id": "",
+    "openmrs_date_concept_id": "",
+    "openmrs_dose_concept_id": ""
+  },
+  {
+    "doses": "10",
+    "name": "Penta",
+    "openmrs_parent_entity_id": "",
+    "openmrs_date_concept_id": "",
+    "openmrs_dose_concept_id": ""
+  },
+  {
+    "doses": "2",
+    "name": "PCV",
+    "openmrs_parent_entity_id": "",
+    "openmrs_date_concept_id": "",
+    "openmrs_dose_concept_id": ""
+  },
+  {
+    "doses": "10",
+    "name": "Rota",
+    "openmrs_parent_entity_id": "",
+    "openmrs_date_concept_id": "",
+    "openmrs_dose_concept_id": ""
+  },
+  {
+    "doses": "10",
+    "name": "M/MR",
+    "openmrs_parent_entity_id": "",
+    "openmrs_date_concept_id": "",
+    "openmrs_dose_concept_id": ""
+  }
+]
