+++ opensrp-path/src/main/java/org/ei/opensrp/path/domain/Vaccine_types.java
@@ -5,14 +5,14 @@ package org.ei.opensrp.path.domain;
  */
 public class Vaccine_types {
 
-    private long id;
+    private Long id;
     private int doses;
     private String name;
     private String openmrs_parent_entity_id;
     private String openmrs_date_concept_id;
     private String openmrs_dose_concept_id;
 
-    public Vaccine_types(long id, int doses, String name, String openmrs_parent_entity_id, String openmrs_date_concept_id, String openmrs_dose_concept_id) {
+    public Vaccine_types(Long id, int doses, String name, String openmrs_parent_entity_id, String openmrs_date_concept_id, String openmrs_dose_concept_id) {
         this.id = id;
         this.doses = doses;
         this.name = name;
