+++ opensrp-path/src/main/res/menu/activity_base_drawer.xml
@@ -10,6 +10,10 @@
             android:id="@+id/nav_record_vaccination_out_catchment"
             android:icon="@drawable/ic_action_out_catchment"
             android:title="Record out of catchment vaccination" />
+        <item
+            android:id="@+id/stock"
+            android:icon="@drawable/ic_reporting"
+            android:title="Stocks" />
         <!--item
             android:id="@+id/nav_settings"
             android:icon="@drawable/ic_action_settings"
